test: rename tests for new namings
