+++ .github/workflows/ci.yml
@@ -286,7 +286,7 @@ jobs:
           if [[ "${{ matrix.test-path }}" =~ ^tests/distributed/* || "${{ matrix.test-path }}" =~ ^tests/daemon/* ]]; then
             # Build daemon for all daemon/distributed tests
             docker build -f Dockerfiles/debianx.Dockerfile --build-arg PIP_TAG=daemon -t jinaai/jina:test-daemon .
-            declare -a distributed_tests=("tests/distributed/test_topologies/" "tests/distributed/test_topologies_docker/" "tests/distributed/test_workspaces/" "tests/distributed/test_dir_structures/" "tests/distributed/test_scale_remote/" "tests/distributed/test_scale_remote_executors/" "tests/distributed/test_env_vars/" "tests/distributed/test_remote_pods/" "tests/distributed/test_remote_flows/" "tests/distributed/test_remote_peas/")
+            declare -a distributed_tests=("tests/distributed/test_topologies/" "tests/distributed/test_topologies_docker/" "tests/distributed/test_workspaces/" "tests/distributed/test_dir_structures/" "tests/distributed/test_scale_remote/" "tests/distributed/test_scale_remote_executors/" "tests/distributed/test_env_vars/" "tests/distributed/test_remote_pods/" "tests/distributed/test_remote_flows/" "tests/distributed/test_remote_deployments/")
             if [[ " ${distributed_tests[*]} " =~ "${{ matrix.test-path }}" ]]; then
               docker run --add-host=host.docker.internal:host-gateway --name jinad --env JINA_DAEMON_BUILD=DEVEL --env JINA_RANDOM_PORT_MIN=12500 --env JINA_RANDOM_PORT_MAX=16383 -v /var/run/docker.sock:/var/run/docker.sock -v /tmp/jinad:/tmp/jinad -p 8000:8000 -d jinaai/jina:test-daemon
             fi
