+++ tests/distributed/test_remote_deployments/test_remote_deployments.py
@@ -1,0 +1,96 @@
+import os
+
+import pytest
+
+from daemon.clients import AsyncJinaDClient, JinaDClient
+from jina import __default_host__
+from jina.enums import replace_enum_to_str
+from jina.helper import ArgNamespace
+from jina.parsers import set_deployment_parser
+
+HOST = __default_host__
+PORT = 8000
+cur_dir = os.path.dirname(os.path.abspath(__file__))
+
+
+@pytest.fixture
+def pod_args():
+    return set_deployment_parser().parse_args([])
+
+
+@pytest.fixture
+def jinad_client():
+    return JinaDClient(host=HOST, port=PORT)
+
+
+@pytest.fixture
+def async_jinad_client():
+    return AsyncJinaDClient(host=HOST, port=PORT)
+
+
+def test_remote_jinad_pod(pod_args, jinad_client):
+    payload = replace_enum_to_str(ArgNamespace.flatten_to_dict(pod_args))
+
+    workspace_id = jinad_client.workspaces.create(
+        paths=[os.path.join(cur_dir, cur_dir)]
+    )
+    assert jinad_client.deployments.alive()
+    # create pod
+    success, pod_id = jinad_client.deployments.create(
+        workspace_id=workspace_id, payload=payload
+    )
+    assert success
+    # get pod status
+    remote_pod_args = jinad_client.deployments.get(pod_id)['arguments']['object'][
+        'arguments'
+    ]
+    # delete pod
+    assert jinad_client.deployments.delete(pod_id)
+    resp = jinad_client.deployments.get(pod_id)
+    assert resp == pod_id + ' not found in store'
+
+
+@pytest.mark.asyncio
+async def test_remote_jinad_pod_async(pod_args, async_jinad_client):
+    payload = replace_enum_to_str(ArgNamespace.flatten_to_dict(pod_args))
+
+    workspace_id = await async_jinad_client.workspaces.create(
+        paths=[os.path.join(cur_dir, cur_dir)]
+    )
+    assert await async_jinad_client.deployments.alive()
+    # create pod
+    success, pod_id = await async_jinad_client.deployments.create(
+        workspace_id=workspace_id, payload=payload
+    )
+    assert success
+    # get pod status
+    resp = await async_jinad_client.deployments.get(pod_id)
+    remote_pod_args = resp['arguments']['object']['arguments']
+    # delete pod
+    assert await async_jinad_client.deployments.delete(pod_id)
+    resp = await async_jinad_client.deployments.get(pod_id)
+    assert resp == pod_id + ' not found in store'
+
+
+@pytest.mark.asyncio
+async def test_jinad_pod_create_async_given_unprocessable_entity(
+    pod_args, async_jinad_client
+):
+    payload = replace_enum_to_str(ArgNamespace.flatten_to_dict(pod_args))
+    payload['pod_role'] = 'RANDOM'  # patch an invalid pod role type
+
+    workspace_id = await async_jinad_client.workspaces.create(
+        paths=[os.path.join(cur_dir, cur_dir)]
+    )
+    success, resp = await async_jinad_client.deployments.create(
+        workspace_id=workspace_id, payload=payload
+    )
+    assert not success
+    assert 'validation error in the payload' in resp
+
+
+@pytest.mark.asyncio
+async def test_jinad_pod_arguments(pod_args, async_jinad_client):
+    resp = await async_jinad_client.deployments.arguments()
+    assert resp
+    assert resp['name']['title'] == 'Name'
